feat(cli): run_with_plugins library entry point
The labwired binary's guts move into the labwired-cli library behind
run_with_plugins(&[&dyn ChipPlugin]); src/main.rs is a thin shim passing
&[]. The version check runs before any command dispatch, `chips` also
lists plugin chip names, and every chip-resolution / bus-construction
path the CLI performs is plugin-aware (resolve_with /
from_config_with_plugins threaded through run, test, machine, snapshot,
debug-probe, asset validate, and environment tests).

Small plugin-aware variants added outside crates/cli, following the
established delegation pattern (old entry points delegate with &[] /
&|_| None):
- labwired_config: ResolvedSystem::from_builtin_chip_with_plugins and
  ResolvedSystem::chip_with_plugins
- labwired_core: system::builder::build_system_bus_with_plugins and
  build_esp32_system_from_manifest_with_plugins,
  system::node::build_node_with_plugins, World::from_manifest_with_plugins
